move cli tools to io folder
